+++ src/components/Button/index.js
@@ -3,6 +3,8 @@ import React, { PureComponent as Component } from 'react'
 import { propConnect } from '../PropProvider'
 import ButtonV2 from './ButtonV2'
 import Button from './Button'
+import { setComponentKey } from '../../utilities/component'
+import { COMPONENT_KEY } from './utils'
 
 type Props = {
   version?: number,
@@ -15,5 +17,7 @@ class WrappedButton extends Component<Props> {
     return version === 2 ? <ButtonV2 {...rest} /> : <Button {...rest} />
   }
 }
+
+setComponentKey(WrappedButton, COMPONENT_KEY)
 
 export default propConnect('Button')(WrappedButton)
