+++ src/components/ControlGroup/styles/ControlGroup.css.js
@@ -1,0 +1,11 @@
+// @flow
+import baseStyles from '../../../styles/resets/baseStyles.css.js'
+import styled from '../../styled'
+
+export const ControlGroupUI = styled('div')`
+  ${baseStyles} display: flex;
+  position: relative;
+  vertical-align: middle;
+`
+
+export default ControlGroupUI
