+++ src/components/ControlGroup/utils.js
@@ -1,0 +1,2 @@
+export const COMPONENT_KEY = 'ControlGroup'
+export const ITEM_COMPONENT_KEY = 'ControlGroup.Item'
