+++ src/components/Input/Input.js
@@ -4,6 +4,7 @@ import type { UISize, UIState } from '../../constants/types'
 import React, { PureComponent as Component } from 'react'
 import { getValidProps } from '@helpscout/react-utils'
 import FormLabelContext from '../FormLabel/Context'
+import AddOn from './AddOn'
 import Backdrop from './Backdrop'
 import Resizer from './Resizer'
 import Static from './Static'
@@ -11,10 +12,10 @@ import HelpText from '../HelpText'
 import Icon from '../Icon'
 import Label from '../Label'
 import { scrollLockY } from '../ScrollLock'
-import Text from '../Text'
 import Tooltip from '../Tooltip'
 import { STATES } from '../../constants/index'
 import classNames from '../../utilities/classNames'
+import { setComponentKey } from '../../utilities/component'
 import { createUniqueIDFactory } from '../../utilities/id'
 import { noop, requestAnimationFrame } from '../../utilities/other'
 import {
@@ -23,6 +24,7 @@ import {
   moveCursorToEnd,
   isTextArea,
 } from './helpers'
+import { COMPONENT_KEY } from './utils'
 
 const uniqueID = createUniqueIDFactory('Input')
 
@@ -46,6 +48,9 @@ type Props = {
   inlineSuffix?: string,
   inputRef: (ref: HTMLElement) => void,
   isFocused: boolean,
+  isFirst: boolean,
+  isNotOnly: boolean,
+  isLast: boolean,
   label: any,
   modalhelpText: string,
   moveCursorToEnd: boolean,
@@ -81,6 +86,7 @@ type Props = {
 
 type State = {
   id: string,
+  isFocused: boolean,
   height: ?number,
   state: ?UIState,
   typingThrottle: ?IntervalID,
@@ -96,6 +102,9 @@ export class Input extends Component<Props, State> {
     forceAutoFocusTimeout: 0,
     inputRef: noop,
     isFocused: false,
+    isFirst: false,
+    isNotOnly: false,
+    isLast: false,
     moveCursorToEnd: false,
     multiline: null,
     offsetAmount: 0,
@@ -119,15 +128,19 @@ export class Input extends Component<Props, State> {
     value: '',
     withTypingEvent: false,
   }
+  static AddOn = AddOn
   static Backdrop = Backdrop
   static Resizer = Resizer
   static Static = Static
   inputNode: InputNode
 
+  static displayName = COMPONENT_KEY
+
   constructor(props: Props) {
     super(props)
     this.state = {
       id: props.id || uniqueID(),
+      isFocused: props.isFocused,
       height: null,
       state: props.state,
       typingThrottle: undefined,
@@ -179,6 +192,11 @@ export class Input extends Component<Props, State> {
 
   forceAutoFocus() {
     const { forceAutoFocusTimeout } = this.props
+
+    this.setState({
+      isFocused: true,
+    })
+
     setTimeout(() => {
       /* istanbul ignore else */
       if (this.inputNode) {
@@ -279,12 +297,22 @@ export class Input extends Component<Props, State> {
     this.scrollToBottom()
   }
 
+  handleOnInputBlur = (event: InputEvent) => {
+    this.setState({
+      isFocused: false,
+    })
+    this.props.onBlur(event)
+  }
+
   handleOnInputFocus = (event: InputEvent) => {
     const { onFocus, removeStateStylesOnFocus } = this.props
     const { state } = this.state
     if (removeStateStylesOnFocus && state) {
       this.setState({ state: null })
     }
+    this.setState({
+      isFocused: true,
+    })
     this.moveCursorToEnd()
     onFocus(event)
   }
@@ -353,38 +381,6 @@ export class Input extends Component<Props, State> {
         <Label className="c-Input__label" for={inputID}>
           {label}
         </Label>
-      )
-    )
-  }
-
-  getPrefixMarkup = () => {
-    const { prefix } = this.props
-
-    return (
-      prefix && (
-        <div className="c-Input__item c-Input__prefix">
-          <div className="c-Input__prefixContent">
-            <Text block shade="muted">
-              {prefix}
-            </Text>
-          </div>
-        </div>
-      )
-    )
-  }
-
-  getSuffixMarkup = () => {
-    const { suffix } = this.props
-
-    return (
-      suffix && (
-        <div className="c-Input__item c-Input__suffix">
-          <div className="c-Input__suffixContent">
-            <Text block shade="muted">
-              {suffix}
-            </Text>
-          </div>
-        </div>
       )
     )
   }
@@ -476,6 +472,9 @@ export class Input extends Component<Props, State> {
       hintText,
       inputRef,
       isFocused,
+      isFirst,
+      isNotOnly,
+      isLast,
       label,
       maxHeight,
       moveCursorToEnd,
@@ -537,7 +536,7 @@ export class Input extends Component<Props, State> {
       ref: this.setInputNodeRef,
       disabled,
       name,
-      onBlur,
+      onBlur: this.handleOnInputBlur,
       onFocus: this.handleOnInputFocus,
       onWheel: this.handleOnWheel,
       placeholder,
@@ -554,6 +553,9 @@ export class Input extends Component<Props, State> {
     const {
       className,
       disabled,
+      isFirst,
+      isNotOnly,
+      isLast,
       maxHeight,
       multiline,
       readOnly,
@@ -562,11 +564,12 @@ export class Input extends Component<Props, State> {
       style: styleProp,
     } = this.props
 
-    const { value, state } = this.state
+    const { isFocused, value, state } = this.state
 
     const componentClassName = classNames(
       'c-Input',
       disabled && 'is-disabled',
+      isFocused && 'is-focused',
       maxHeight && 'has-maxHeight',
       multiline && 'is-multiline',
       readOnly && 'is-readonly',
@@ -580,8 +583,6 @@ export class Input extends Component<Props, State> {
     const helpTextMarkup = this.getHelpTextMarkup()
     const hintTextMarkup = this.getHintTextMarkup()
     const labelMarkup = this.getLabelMarkup()
-    const prefixMarkup = this.getPrefixMarkup()
-    const suffixMarkup = this.getSuffixMarkup()
     const inlinePrefixMarkup = this.getInlinePrefixMarkup()
     const inlineSuffixMarkup = this.getInlineSuffixMarkup()
     const errorMarkup = this.getErrorMarkup()
@@ -595,15 +596,16 @@ export class Input extends Component<Props, State> {
             {labelMarkup}
             {hintTextMarkup}
             <div className={componentClassName}>
-              {prefixMarkup}
               {inlinePrefixMarkup}
               {this.getInputMarkup(props)}
               {inlineSuffixMarkup}
-              {suffixMarkup}
               {errorMarkup}
               <Backdrop
                 className="c-Input__backdrop"
                 disabled={disabled}
+                isFirst={isFirst}
+                isNotOnly={isNotOnly}
+                isLast={isLast}
                 readOnly={readOnly}
                 state={state}
               />
@@ -616,5 +618,7 @@ export class Input extends Component<Props, State> {
     )
   }
 }
+
+setComponentKey(Input, COMPONENT_KEY)
 
 export default Input
