+++ src/components/Input/utils.js
@@ -1,0 +1,3 @@
+export const COMPONENT_KEY = 'Input'
+export const ADDON_COMPONENT_KEY = 'Input.AddOn'
+export const STATIC_COMPONENT_KEY = 'Input.Static'
