+++ src/components/index.js
@@ -21,7 +21,7 @@ export { default as Choice } from './Choice'
 export { default as ChoiceGroup } from './ChoiceGroup'
 export { default as CloseButton } from './CloseButton'
 export { default as Code } from './Code'
-export { default as PropProvider } from './PropProvider'
+export { default as ControlGroup } from './ControlGroup'
 export { default as Collapsible } from './Collapsible'
 export { default as DetailList } from './DetailList'
 export { default as Dropdown } from './Dropdown'
@@ -61,6 +61,7 @@ export { default as Portal } from './Portal'
 export { default as PortalWrapper } from './PortalWrapper'
 export { default as PreviewCard } from './PreviewCard'
 export { default as ProgressBar } from './ProgressBar'
+export { default as PropProvider } from './PropProvider'
 export { default as RouteWrapper } from './RouteWrapper'
 export { default as Radio } from './Radio'
 export { default as Samp } from './Samp'
