+++ src/styles/mixins/input-styles.scss
@@ -18,38 +18,6 @@
     z-index: 1;
   }
 
-  &__prefix,
-  &__suffix {
-    align-self: stretch;
-    align-items: center;
-    background-color: _color(grey, 200);
-    display: flex;
-    margin: 0;
-    padding: 10px;
-  }
-
-  &__prefix {
-    border-right: 1px solid _color(grey, 600);
-    border-top-left-radius: 4px;
-    border-bottom-left-radius: 4px;
-    margin-left: -15px;
-    margin-right: 13px;
-    padding-left: 14px;
-  }
-  &__suffix {
-    border-left: 1px solid _color(grey, 600);
-    border-top-right-radius: 4px;
-    border-bottom-right-radius: 4px;
-    margin-left: 13px;
-    margin-right: -15px;
-    padding-right: 14px;
-  }
-
-  &__prefixContent,
-  &__suffixContent {
-    display: block;
-  }
-
   &__inlinePrefix,
   &__inlineSuffix {
     opacity: 0.3;
