+++ stories/index.js
@@ -22,6 +22,7 @@ import './ChatTranscript'
 import './Checkbox'
 import './Code'
 import './Collapsible'
+import './ControlGroup'
 import './CloseButton'
 import './DetailList'
 import './Dropdown'
